final-task-django/djangoTeam-tasks-master/djangoTeam-tasks-master/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ninja==0.20.0
pydantic==1.10.5
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
